Extract deleteRecording logic, fix formatting
