+++ src/main/java/io/cryostat/net/web/http/api/v1/TargetRecordingDeleteHandler.java
@@ -38,20 +38,16 @@
 package io.cryostat.net.web.http.api.v1;
 
 import java.util.Map;
-import java.util.Optional;
 
 import javax.inject.Inject;
-
-import org.openjdk.jmc.rjmx.services.jfr.IRecordingDescriptor;
 
 import io.cryostat.messaging.notifications.NotificationFactory;
 import io.cryostat.net.AuthManager;
 import io.cryostat.net.ConnectionDescriptor;
-import io.cryostat.net.TargetConnectionManager;
-import io.cryostat.net.reports.ReportService;
 import io.cryostat.net.web.http.AbstractAuthenticatedRequestHandler;
 import io.cryostat.net.web.http.HttpMimeType;
 import io.cryostat.net.web.http.api.ApiVersion;
+import io.cryostat.recordings.RecordingHelper;
 
 import io.vertx.core.http.HttpMethod;
 import io.vertx.ext.web.RoutingContext;
@@ -59,21 +55,18 @@ import io.vertx.ext.web.handler.impl.HttpStatusException;
 
 class TargetRecordingDeleteHandler extends AbstractAuthenticatedRequestHandler {
 
-    private final TargetConnectionManager targetConnectionManager;
-    private final ReportService reportService;
     private final NotificationFactory notificationFactory;
     private static final String NOTIFICATION_CATEGORY = "RecordingDeleted";
+    private final RecordingHelper recordingHelper;
 
     @Inject
     TargetRecordingDeleteHandler(
             AuthManager auth,
-            TargetConnectionManager targetConnectionManager,
             NotificationFactory notificationFactory,
-            ReportService reportService) {
+            RecordingHelper recordingHelper) {
         super(auth);
         this.notificationFactory = notificationFactory;
-        this.targetConnectionManager = targetConnectionManager;
-        this.reportService = reportService;
+        this.recordingHelper = recordingHelper;
     }
 
     @Override
@@ -100,37 +93,26 @@ class TargetRecordingDeleteHandler extends AbstractAuthenticatedRequestHandler {
     public void handleAuthenticated(RoutingContext ctx) throws Exception {
         String recordingName = ctx.pathParam("recordingName");
         ConnectionDescriptor connectionDescriptor = getConnectionDescriptorFromContext(ctx);
-        targetConnectionManager.executeConnectedTask(
-                connectionDescriptor,
-                connection -> {
-                    Optional<IRecordingDescriptor> descriptor =
-                            connection.getService().getAvailableRecordings().stream()
-                                    .filter(recording -> recording.getName().equals(recordingName))
-                                    .findFirst();
-                    if (descriptor.isPresent()) {
-                        connection.getService().close(descriptor.get());
-                        reportService.delete(connectionDescriptor, recordingName);
-                        notificationFactory
-                                .createBuilder()
-                                .metaCategory(NOTIFICATION_CATEGORY)
-                                .metaType(HttpMimeType.JSON)
-                                .message(
-                                        Map.of(
-                                                "recording",
-                                                recordingName,
-                                                "target",
-                                                connectionDescriptor.getTargetId()))
-                                .build()
-                                .send();
-                        ctx.response().setStatusCode(200);
-                        ctx.response().end();
-                    } else {
-                        throw new HttpStatusException(
-                                404,
-                                String.format(
-                                        "No recording with name \"%s\" found", recordingName));
-                    }
-                    return null;
-                });
+
+        try {
+            recordingHelper.deleteRecording(connectionDescriptor, recordingName);
+        } catch (Exception e) {
+            throw new HttpStatusException(500, e);
+        }
+
+        notificationFactory
+                .createBuilder()
+                .metaCategory(NOTIFICATION_CATEGORY)
+                .metaType(HttpMimeType.JSON)
+                .message(
+                        Map.of(
+                                "recording",
+                                recordingName,
+                                "target",
+                                connectionDescriptor.getTargetId()))
+                .build()
+                .send();
+        ctx.response().setStatusCode(200);
+        ctx.response().end();
     }
 }
