+++ src/test/java/io/cryostat/net/web/http/api/v1/TargetRecordingDeleteHandlerTest.java
@@ -39,41 +39,33 @@ package io.cryostat.net.web.http.api.v1;
 
 import static org.mockito.Mockito.lenient;
 
-import java.util.List;
 import java.util.Map;
 
-import org.openjdk.jmc.common.unit.IQuantity;
-import org.openjdk.jmc.common.unit.QuantityConversionException;
 import org.openjdk.jmc.rjmx.services.jfr.IFlightRecorderService;
-import org.openjdk.jmc.rjmx.services.jfr.IRecordingDescriptor;
 
 import io.cryostat.core.net.JFRConnection;
 import io.cryostat.messaging.notifications.Notification;
 import io.cryostat.messaging.notifications.NotificationFactory;
 import io.cryostat.net.AuthManager;
-import io.cryostat.net.ConnectionDescriptor;
 import io.cryostat.net.TargetConnectionManager;
 import io.cryostat.net.reports.ReportService;
 import io.cryostat.net.web.http.HttpMimeType;
+import io.cryostat.recordings.RecordingHelper;
 
 import io.vertx.core.MultiMap;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.http.HttpServerResponse;
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.handler.impl.HttpStatusException;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InOrder;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.invocation.InvocationOnMock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.mockito.stubbing.Answer;
 
 @ExtendWith(MockitoExtension.class)
 class TargetRecordingDeleteHandlerTest {
@@ -85,6 +77,7 @@ class TargetRecordingDeleteHandlerTest {
     @Mock Notification notification;
     @Mock Notification.Builder notificationBuilder;
     @Mock ReportService reportService;
+    @Mock RecordingHelper recordingHelper;
 
     @Mock RoutingContext ctx;
     @Mock HttpServerRequest req;
@@ -106,9 +99,7 @@ class TargetRecordingDeleteHandlerTest {
                 .thenReturn(notificationBuilder);
         lenient().when(notificationBuilder.message(Mockito.any())).thenReturn(notificationBuilder);
         lenient().when(notificationBuilder.build()).thenReturn(notification);
-        this.handler =
-                new TargetRecordingDeleteHandler(
-                        auth, targetConnectionManager, notificationFactory, reportService);
+        this.handler = new TargetRecordingDeleteHandler(auth, notificationFactory, recordingHelper);
     }
 
     @Test
@@ -130,33 +121,9 @@ class TargetRecordingDeleteHandlerTest {
         Mockito.when(ctx.request()).thenReturn(req);
         Mockito.when(ctx.response()).thenReturn(resp);
         Mockito.when(ctx.request().headers()).thenReturn(MultiMap.caseInsensitiveMultiMap());
-        Mockito.when(targetConnectionManager.executeConnectedTask(Mockito.any(), Mockito.any()))
-                .thenAnswer(
-                        new Answer() {
-                            @Override
-                            public Object answer(InvocationOnMock invocation) throws Throwable {
-                                TargetConnectionManager.ConnectedTask task =
-                                        (TargetConnectionManager.ConnectedTask)
-                                                invocation.getArgument(1);
-                                return task.execute(connection);
-                            }
-                        });
-
-        Mockito.when(connection.getService()).thenReturn(service);
-        IRecordingDescriptor descriptor = createDescriptor("someRecording");
-        Mockito.when(service.getAvailableRecordings()).thenReturn(List.of(descriptor));
 
         handler.handleAuthenticated(ctx);
 
-        Mockito.verify(service).close(descriptor);
-        ConnectionDescriptor connectionDescriptor = new ConnectionDescriptor("fooTarget");
-        Mockito.verify(reportService)
-                .delete(
-                        Mockito.argThat(
-                                arg ->
-                                        arg.getTargetId()
-                                                .equals(connectionDescriptor.getTargetId())),
-                        Mockito.eq("someRecording"));
         InOrder inOrder = Mockito.inOrder(resp);
         inOrder.verify(resp).setStatusCode(200);
         inOrder.verify(resp).end();
@@ -168,50 +135,5 @@ class TargetRecordingDeleteHandlerTest {
                 .message(Map.of("recording", "someRecording", "target", "fooTarget"));
         Mockito.verify(notificationBuilder).build();
         Mockito.verify(notification).send();
-    }
-
-    @Test
-    void shouldHandleRecordingNotFound() throws Exception {
-        Mockito.when(ctx.request()).thenReturn(req);
-        Mockito.when(ctx.request().headers()).thenReturn(MultiMap.caseInsensitiveMultiMap());
-        Mockito.when(ctx.pathParam("targetId")).thenReturn("fooTarget");
-        Mockito.when(ctx.pathParam("recordingName")).thenReturn("someRecording");
-        Mockito.when(targetConnectionManager.executeConnectedTask(Mockito.any(), Mockito.any()))
-                .thenAnswer(
-                        new Answer() {
-                            @Override
-                            public Object answer(InvocationOnMock invocation) throws Throwable {
-                                TargetConnectionManager.ConnectedTask task =
-                                        (TargetConnectionManager.ConnectedTask)
-                                                invocation.getArgument(1);
-                                return task.execute(connection);
-                            }
-                        });
-
-        Mockito.when(connection.getService()).thenReturn(service);
-        Mockito.when(service.getAvailableRecordings()).thenReturn(List.of());
-
-        HttpStatusException ex =
-                Assertions.assertThrows(
-                        HttpStatusException.class, () -> handler.handleAuthenticated(ctx));
-        MatcherAssert.assertThat(ex.getStatusCode(), Matchers.equalTo(404));
-    }
-
-    private static IRecordingDescriptor createDescriptor(String name)
-            throws QuantityConversionException {
-        IQuantity zeroQuantity = Mockito.mock(IQuantity.class);
-        IRecordingDescriptor descriptor = Mockito.mock(IRecordingDescriptor.class);
-        Mockito.lenient().when(descriptor.getId()).thenReturn(1L);
-        Mockito.lenient().when(descriptor.getName()).thenReturn(name);
-        Mockito.lenient()
-                .when(descriptor.getState())
-                .thenReturn(IRecordingDescriptor.RecordingState.STOPPED);
-        Mockito.lenient().when(descriptor.getStartTime()).thenReturn(zeroQuantity);
-        Mockito.lenient().when(descriptor.getDuration()).thenReturn(zeroQuantity);
-        Mockito.lenient().when(descriptor.isContinuous()).thenReturn(false);
-        Mockito.lenient().when(descriptor.getToDisk()).thenReturn(false);
-        Mockito.lenient().when(descriptor.getMaxSize()).thenReturn(zeroQuantity);
-        Mockito.lenient().when(descriptor.getMaxAge()).thenReturn(zeroQuantity);
-        return descriptor;
     }
 }
